c-mandelbrot-koodissa kaikki doublet vaihdettu floateiksi; kommentteja

+++ trunk/fractal.h
@@ -23,8 +23,8 @@ typedef struct {
 
 
 void drawMandelbrotArea( uint32 width, uint32 height,
-                         double reOffset, double imOffset,
-                         double zoom, uint32 maxIteration,
+                         float reOffset, float imOffset,
+                         float zoom, uint32 maxIteration,
                          char *areaBuffer,
                          uint32 areaX, uint32 areaY,
                          uint32 areaWidth, uint32 areaHeight,
